models/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.9
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.231
langchainplus-sdk==0.0.20
llama-index==0.7.5
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
